Minor renames for controller
