+++ src/test/java/uk/gov/hmcts/reform/ccd/document/am/WelcomeTest.java
@@ -3,7 +3,7 @@ package uk.gov.hmcts.reform.ccd.document.am;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import uk.gov.hmcts.reform.ccd.document.am.controllers.CaseDocumentAmController;
+import uk.gov.hmcts.reform.ccd.document.am.controller.CaseDocumentAmController;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
